refactor: Move components separate directory
