+++ src/components/Bookmarks.test.ts
@@ -1,7 +1,7 @@
 // @vitest-environment jsdom
 import { render } from '@testing-library/svelte';
 import { describe, expect, it } from 'vitest';
-import Bookmarks from '~/pages/options/Bookmarks.svelte';
+import Bookmarks from '~/components/Bookmarks.svelte';
 
 describe('ResourceView', () => {
 	it('has some descriptive heading', () => {
