+++ src/components/Config.svelte
